int64 -> uint

+++ internal/model/category.go
@@ -3,7 +3,7 @@ package model
 type Category struct {
 	Model
 	Title   string `gorm:"column:title;not null"`
-	StoreID int64  `gorm:"column:store_id"`
+	StoreID uint   `gorm:"column:store_id"`
 
 	Store *Store
 }
