+++ internal/model/customer.go
@@ -2,7 +2,7 @@ package model
 
 type Customer struct {
 	Model
-	ExternalID int64  `gorm:"column:external_id;not null"`
+	ExternalID uint   `gorm:"column:external_id;not null"`
 	Username   string `gorm:"column:username"`
 	Phone      string `gorm:"column:phone"`
 	FirstName  string `gorm:"column:first_name"`
