+++ internal/model/order.go
@@ -8,8 +8,8 @@ type Order struct {
 	State              enums.OrderState    `gorm:"column:state;not null"`
 	PaymentMethod      enums.PaymentMethod `gorm:"column:payment_method;not null"`
 	CancellationReason string              `gorm:"column:cancellation_reason"`
-	CustomerID         int64               `gorm:"column:customer_id"`
-	StoreID            int64               `gorm:"column:store_id"`
+	CustomerID         uint                `gorm:"column:customer_id"`
+	StoreID            uint                `gorm:"column:store_id"`
 
 	Customer *Customer
 	Store    *Store
