+++ internal/model/order_item.go
@@ -1,10 +1,10 @@
 package model
 
 type OrderItem struct {
-	ID        int64 `gorm:"column:id"`
-	Amount    int   `gorm:"column:amount;not null;default:1;check:amount > 0"`
-	OrderID   int64 `gorm:"column:order_id"`
-	ProductID int64 `gorm:"column:product_id"`
+	ID        uint `gorm:"column:id"`
+	Amount    int  `gorm:"column:amount;not null;default:1;check:amount > 0"`
+	OrderID   uint `gorm:"column:order_id"`
+	ProductID uint `gorm:"column:product_id"`
 
 	Order   *Order
 	Product *Product
